refactor: move text constants to shared/lib/locales
