+++ src/pages/landing/ui/training/index.tsx
@@ -11,7 +11,7 @@ import { BUTTON_TEXT, LANDING_HERO_TEXT, LANDING_SUCCESS_TEXT } from '../../loca
 import styles from './styles.module.css';
 
 import { ROUTES, Button, cn } from '@/shared';
-import { HEADERS_TEXT } from '@/shared/model/constants';
+import { HEADERS_TEXT } from '@/shared/lib/locales';
 
 export const Training = () => {
   const { ref, isVisible } = useIntersectionObserver({
