--- src/shared/model/constants.ts
@@ -1,3 +1,0 @@
-export const HEADERS_TEXT = {
-  SITE_TITLE: 'Тренажёр интервью',
-};
